+++ crates/nostr-sdk/CHANGELOG.md
@@ -29,6 +29,10 @@
 
 - Replace `usize` with `u8` for gossip relay limits
 
+### Added
+
+- Add `GossipAllowedRelays` to `GossipOptions` to filter relays during selection (https://github.com/rust-nostr/nostr/pull/1128)
+
 ## v0.44.1 - 2025/11/09
 
 ### Fixed